Create static LoginDetails class
Allows me to store login details in memory.
This allows the users details to persist through page changes.

The user still has to login again but the details are filled out for them.
User has to login again so that they're able to change login info and API urls if needed, without restarting the whole Ui.